# app/swagger.yaml
swagger: "2.0"
info:
  title: "Banking API"
  description: "API documentation for the Banking application."
  version: "1.0.0"
host: "localhost:5000"
basePath: "/api"
schemes:
  - "http"
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter your JWT token in the format **Bearer &lt;token&gt;**"
paths:
...
  /accounts:
    get:
      summary: "Get User Accounts"
      description: "Retrieve a list of accounts for the authenticated user."
      tags:
        - "Accounts"
      security:
        - Bearer: []
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          description: "Page number for pagination"
          example: 1
        - in: "query"
          name: "per_page"
          type: "integer"
          description: "Number of accounts per page"
          example: 10
        - in: "query"
          name: "type"
          type: "string"
          description: "Filter accounts by type (e.g., savings, checking)"
          example: "savings"
      responses:
        200:
          description: "List of user accounts"
          schema:
            type: "object"
            properties:
              account_listing:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "integer"
                      example: 1
                    category:
                      type: "string"
                      example: "savings"
                    label:
                      type: "string"
                      example: "My Savings Account"
                    balance:
                      type: "number"
                      format: "float"
                      example: 1000.5
              page:
                type: "integer"
                example: 1
              per_page:
                type: "integer"
                example: 10
              total:
                type: "integer"
                example: 2

    post:
      summary: "Create a New Account"
      description: "Create a new account for the authenticated user."
      tags:
        - "Accounts"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Account creation details"
          schema:
            type: "object"
            required:
              - account_type
              - account_name
            properties:
              account_type:
                type: "string"
                example: "savings"
              account_name:
                type: "string"
                example: "My Savings Account"
              initial_balance:
                type: "number"
                format: "float"
                example: 500.0
              description:
                type: "string"
                example: "This is my savings account."
      responses:
        201:
          description: "Account created successfully"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                example: 1
              category:
                type: "string"
                example: "savings"
              label:
                type: "string"
                example: "My Savings Account"
              balance:
                type: "number"
                format: "float"
                example: 500.0
              message:
                type: "string"
                example: "Account created successfully"
  /accounts/{account_id}:
    get:
      summary: "Get Account Details"
      description: "Retrieve details of a specific account by its ID."
      tags:
        - "Accounts"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to retrieve"
      responses:
        200:
          description: "Account details retrieved successfully"
          schema:
            type: "object"
            properties:
              account_detail:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  category:
                    type: "string"
                    example: "savings"
                  label:
                    type: "string"
                    example: "My Savings Account"
                  balance:
                    type: "number"
                    format: "float"
                    example: 1000.5
              balance:
                type: "number"
                format: "float"
                example: 1000.5
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found"

    put:
      summary: "Update Account"
      description: "Update details of a specific account by its ID."
      tags:
        - "Accounts"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to update"
        - in: "body"
          name: "body"
          required: true
          description: "Account update details"
          schema:
            type: "object"
            properties:
              account_label:
                type: "string"
                example: "Updated Account Name"
              description:
                type: "string"
                example: "Updated account description."
      responses:
        200:
          description: "Account updated successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account updated successfully"
              account_detail:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  category:
                    type: "string"
                    example: "savings"
                  label:
                    type: "string"
                    example: "Updated Account Name"
                  balance:
                    type: "number"
                    format: "float"
                    example: 1000.5
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found or access denied"

    delete:
      summary: "Delete Account"
      description: "Soft delete an account by its ID."
      tags:
        - "Accounts"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to delete"
      responses:
        200:
          description: "Account deletion processed"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account deletion processed"
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found or access denied"
  /accounts/{account_id}/transactions:
    get:
      summary: "Get Account Transactions"
      description: "Retrieve a list of transactions for a specific account."
      tags:
        - "Accounts"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to retrieve transactions for"
        - in: "query"
          name: "start_date"
          type: "string"
          format: "date"
          description: "Filter transactions from this date (YYYY-MM-DD)"
          example: "2023-01-01"
        - in: "query"
          name: "end_date"
          type: "string"
          format: "date"
          description: "Filter transactions up to this date (YYYY-MM-DD)"
          example: "2023-12-31"
        - in: "query"
          name: "type"
          type: "string"
          description: "Filter transactions by type (deposit, withdrawal, transfer)"
          example: "deposit"
        - in: "query"
          name: "search"
          type: "string"
          description: "Search transactions by description"
          example: "groceries"
        - in: "query"
          name: "page"
          type: "integer"
          description: "Page number for pagination"
          example: 1
        - in: "query"
          name: "per_page"
          type: "integer"
          description: "Number of transactions per page"
          example: 20
      responses:
        200:
          description: "List of transactions"
          schema:
            type: "object"
            properties:
              transactions:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "integer"
                      example: 1
                    description:
                      type: "string"
                      example: "Grocery shopping"
                    amount:
                      type: "number"
                      format: "float"
                      example: 50.0
                    timestamp:
                      type: "string"
                      format: "date-time"
                      example: "2023-01-01T12:00:00Z"
              pg:
                type: "integer"
                example: 1
              per_pg:
                type: "integer"
                example: 20
              total_items:
                type: "integer"
                example: 100
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found"
...
  /transactions/accounts/{account_id}/transactions:
    get:
      summary: "Get Transactions for a Specific Account"
      description: "Retrieve a list of transactions for a specific account."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to retrieve transactions for"
      responses:
        200:
          description: "List of transactions for the account"
          schema:
            type: "object"
            properties:
              transactions:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "integer"
                      example: 1
                    transaction_type:
                      type: "string"
                      example: "deposit"
                    amount:
                      type: "number"
                      format: "float"
                      example: 100.0
                    timestamp:
                      type: "string"
                      format: "date-time"
                      example: "2023-01-01T12:00:00Z"
                    description:
                      type: "string"
                      example: "Deposit to savings account"
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found or does not belong to you"

    post:
      summary: "Create a Transaction for a Specific Account"
      description: "Create a new transaction (deposit, withdrawal, or transfer) for a specific account."
      tags:
        - "Transactions"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "account_id"
          required: true
          type: "integer"
          description: "ID of the account to create a transaction for"
        - in: "body"
          name: "body"
          required: true
          description: "Details of the transaction"
          schema:
            type: "object"
            required:
              - type
              - amount
            properties:
              type:
                type: "string"
                example: "deposit"
                description: "Type of transaction (deposit, withdrawal, transfer)"
              amount:
                type: "number"
                format: "float"
                example: 200.0
              to_account_id:
                type: "integer"
                example: 2
                description: "Required for transfers"
              description:
                type: "string"
                example: "Deposit to savings account"
      responses:
        201:
          description: "Transaction created successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Deposit successful"
              transaction:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  transaction_type:
                    type: "string"
                    example: "deposit"
                  amount:
                    type: "number"
                    format: "float"
                    example: 200.0
                  timestamp:
                    type: "string"
                    format: "date-time"
                    example: "2023-01-01T12:00:00Z"
              new_balance:
                type: "number"
                format: "float"
                example: 1200.0
        400:
          description: "Invalid request or insufficient funds"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Invalid transaction type or insufficient funds"
        404:
          description: "Account not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Account not found or does not belong to you"